created meal detail page

+++ data/src/main/java/com/gkgio/data/cookers/CookersDataWithoutAuthRequest.kt
@@ -6,9 +6,9 @@ import com.squareup.moshi.JsonClass
 @JsonClass(generateAdapter = true)
 data class CookersDataWithoutAuthRequest(
     @Json(name = "lat")
-    val lat: Double,
+    val lat: Int,
     @Json(name = "lon")
-    val lon: Double,
+    val lon: Int,
     @Json(name = "distance")
     val distance: String?,
     @Json(name = "targets")
